+++ integration_tests/conftest.py
@@ -1,15 +1,19 @@
 import pytest
+import pytest_asyncio
 import redis
 from redis.asyncio import Redis as RedisAsync
 
 from py_cachify import Cachify, init_cachify
 
 
-@pytest.fixture(autouse=True)
-def init_cachify_fixture() -> None:
+@pytest_asyncio.fixture(autouse=True)  # pyright: ignore[reportUnknownMemberType]
+async def init_cachify_fixture() -> None:
+    sync_client = redis.Redis.from_url(url='redis://localhost:6379/0')  # pyright: ignore[reportUnknownMemberType]
+    async_client = RedisAsync.from_url(url='redis://localhost:6379/1')  # pyright: ignore[reportUnknownMemberType]
+
     init_cachify(
-        sync_client=redis.Redis.from_url(url='redis://localhost:6379/0'),  # pyright: ignore[reportUnknownMemberType]
-        async_client=RedisAsync.from_url(url='redis://localhost:6379/1'),  # pyright: ignore[reportUnknownMemberType]
+        sync_client=sync_client,
+        async_client=async_client,
     )
 
 
@@ -20,10 +24,13 @@ def cachify_local_in_memory_client() -> Cachify:
     )
 
 
-@pytest.fixture
-def cachify_local_redis_second() -> Cachify:
+@pytest_asyncio.fixture  # pyright: ignore[reportUnknownMemberType]
+async def cachify_local_redis_second() -> Cachify:
+    async_client = RedisAsync.from_url(url='redis://localhost:6379/3')  # pyright: ignore[reportUnknownMemberType]
+    sync_client = redis.Redis.from_url(url='redis://localhost:6379/2')  # pyright: ignore[reportUnknownMemberType]
+
     return init_cachify(
         is_global=False,
-        sync_client=redis.Redis.from_url(url='redis://localhost:6379/2'),  # pyright: ignore[reportUnknownMemberType]
-        async_client=RedisAsync.from_url(url='redis://localhost:6379/3'),  # pyright: ignore[reportUnknownMemberType]
+        sync_client=sync_client,
+        async_client=async_client,
     )